# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "description": "Authentifie un membre et retourne un JWT et un refresh token. Requiert que l'email soit vérifié.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connexion réussie — retourne le profil, les capacités et les tokens"
          },
          "401": {
            "description": "Identifiants invalides, email non vérifié, ou compte suspendu"
          },
          "429": {
            "description": "Trop de tentatives de connexion"
          }
        },
        "summary": "Connexion membre",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "description": "Fonctionne pour un membre comme pour un administrateur.",
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profil courant"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Profil de l’acteur connecté",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Journalise la déconnexion. L'authentification étant sans état, le client doit supprimer ses tokens ; la révocation côté serveur nécessiterait une table de sessions.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Déconnexion enregistrée"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Déconnexion",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh-token": {
      "get": {
        "description": "Génère une nouvelle paire de tokens à partir du refresh token fourni dans le header Authorization.",
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Nouvelle paire de tokens"
          },
          "401": {
            "description": "Refresh token invalide ou expiré"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Rafraîchissement du token",
        "tags": [
          "Auth"
        ]
      },
      "post": {
        "operationId": "AuthController_refreshTokenPost",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Nouvelle paire de tokens"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Rafraîchissement du token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "description": "Crée un compte non vérifié et envoie un code OTP à l'adresse email fournie.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Compte créé — un code OTP a été envoyé à l'email",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Email déjà utilisé ou données invalides"
          },
          "429": {
            "description": "Trop de créations de compte"
          }
        },
        "summary": "Création de compte membre",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify-email": {
      "post": {
        "description": "Vérifie le code OTP reçu par email. En cas de succès, retourne les tokens d'authentification.",
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email vérifié — retourne le profil et les tokens"
          },
          "400": {
            "description": "Code OTP invalide ou expiré"
          }
        },
        "summary": "Vérification email par OTP",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/resend-email-otp": {
      "post": {
        "description": "Génère un nouveau code OTP. Réponse générique : ne révèle pas l'existence du compte.",
        "operationId": "AuthController_resendEmailOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendEmailOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Demande prise en compte",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Trop d'envois de code"
          }
        },
        "summary": "Renvoi du code OTP",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "description": "Envoie un code de réinitialisation. Retourne toujours un message générique pour ne pas révéler l'existence du compte.",
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Demande prise en compte",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Mot de passe oublié",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "description": "Définit un nouveau mot de passe à partir du code de réinitialisation reçu par email.",
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mot de passe réinitialisé",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Code de réinitialisation invalide ou expiré"
          }
        },
        "summary": "Réinitialisation du mot de passe",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Citoyen+ API",
    "description": "API officielle de la plateforme Citoyen+ — application citoyenne de la Côte d'Ivoire. Permet la gestion des actualités, signalements citoyens, bibliothèque de documents et quiz éducatifs.",
    "version": "1.0",
    "contact": {
      "name": "Équipe Citoyen+",
      "url": "https://mec-ci.org",
      "email": "contact@mec-ci.org"
    }
  },
...
  "servers": [
    {
      "url": "https://api.mec-ci.org",
      "description": "Production"
    },
    {
      "url": "http://localhost:8081",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "LoginUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email du compte membre",
            "example": "jean.dupont@example.com",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "description": "Mot de passe",
            "example": "MonMotDePasse!2026"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RegisterUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email de l'utilisateur",
            "example": "jean.dupont@example.com",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "description": "Mot de passe de l'utilisateur",
            "example": "MonMotDePasse!2026",
            "minLength": 12,
            "maxLength": 128
          },
          "fullname": {
            "type": "string",
            "description": "nom complet de l'utilisateur",
            "example": "Jean Dupont",
            "maxLength": 100
          },
          "phone": {
            "type": "string",
            "description": "numéro de téléphone de l'utilisateur",
            "example": "+2250701020304",
            "maxLength": 20
          }
        },
        "required": [
          "email",
          "password",
          "fullname"
        ]
      },
      "VerifyEmailOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Adresse email de l'utilisateur",
            "example": "jean.dupont@example.com",
            "maxLength": 254
          },
          "otp": {
            "type": "string",
            "description": "Code OTP à 6 chiffres reçu par email",
            "example": "483921",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "email",
          "otp"
        ]
      },
      "ResendEmailOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Adresse email du compte en attente de vérification",
            "example": "jean.dupont@example.com",
            "maxLength": 254
          }
        },
        "required": [
          "email"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Adresse email du compte à réinitialiser",
            "example": "jean.dupont@example.com",
            "maxLength": 254
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Adresse email du compte à réinitialiser",
            "example": "jean.dupont@example.com",
            "maxLength": 254
          },
          "otp": {
            "type": "string",
            "description": "Code de réinitialisation reçu par email (6 chiffres)",
            "example": "123456"
          },
          "password": {
            "type": "string",
            "description": "Nouveau mot de passe",
            "example": "MonMotDePasse!2026",
            "minLength": 12,
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "otp",
          "password"
        ]
      },
...
    }
  }
}